pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "opticam"
version = "0.4.0"
description = "A Python package for reducing OPTICAM data."
readme = "README.md"
authors = [
    {name = "Zackery A. Irving", email = "zackery.irving@icloud.com"},
]

requires-python = ">=3.12.7"
dependencies = [
    "astroalign == 2.6.0",
    "astropy == 6.1.4",
    "ccdproc == 2.4.2",
    "matplotlib == 3.9.2",
    "numpy == 2.1.3",
    "pandas == 2.2.3",
    "pillow == 11.0.0",
    "photutils == 2.0.2",
    "scikit-image == 0.24.0",
    "scipy == 1.14.1",
    "stingray == 2.2.6",
    "tqdm == 4.66.6",
]

[project.optional-dependencies]
docs = [
    "jupyter >=1.1.1",
    "nbsphinx >= 0.9",
    "sphinx >= 8.1",
    "sphinx-autoapi >= 3.4",
    "sphinx-rtd-theme >= 3.0",
]

[tool.setuptools.packages.find]
where =["."]